refactor zenNode logic to own file
