+++ app/main.dev.js
@@ -12,10 +12,9 @@
  */
 import path from 'path'
 
-import { app, BrowserWindow, ipcMain } from 'electron'
-import zenNode from '@zen/zen-node'
-import _ from 'lodash'
+import { app, BrowserWindow } from 'electron'
 
+import zenNode from './utils/zenNode'
 import db from './services/store'
 import MenuBuilder from './menu'
 
@@ -55,13 +54,6 @@ const installExtensions = async () => {
  * Add event listeners...
  */
 
-let node = {
-  stderr: { pipe: _.noop },
-  stdout: { pipe: _.noop, on: _.noop },
-  on: _.noop,
-  kill: _.noop,
-}
-
 app.on('ready', async () => {
   if (process.env.NODE_ENV === 'development' || process.env.DEBUG_PROD === 'true') {
     await installExtensions()
@@ -75,19 +67,15 @@ app.on('ready', async () => {
   mainWindow.on('resize', saveWindowDimensionsToDb)
 
   if (!isUiOnly) {
-    console.log('LAUNCHING ZEN NODE')
-    try {
-      node = zenNode(getZenNodeArgs(), getZenNodePath())
-      node.stderr.pipe(process.stderr)
-      node.stdout.pipe(process.stdout)
-      setupLogFetching()
-      node.on('exit', () => {
+    zenNode.init()
+    zenNode.node.on('exit', (code, signal) => {
+      if (signal === 'SIGUSR1') {
+        console.log('toggle minning through GUI')
+      } else {
         console.log('Closed')
         app.quit()
-      })
-    } catch (err) {
-      console.error('error launching zen node', err.message, err)
-    }
+      }
+    })
   }
 
   mainWindow.loadURL(`file://${__dirname}/app.html`)
@@ -119,26 +107,10 @@ app.on('window-all-closed', () => {
 
 app.on('will-quit', () => {
   if (!isUiOnly) {
-    console.log('Gracefully shutting down the zen-node after receiving (will-quit)')
-    node.kill('SIGINT')
+    console.log('Gracefully shutting down the zen-node by clicking the close button')
+    zenNode.node.kill('SIGINT')
   }
 })
-
-function getZenNodePath() {
-  return isInstalledWithInstaller()
-    // $FlowFixMe
-    ? path.join(process.resourcesPath, '/node_modules/@zen/zen-node')
-    : path.join(__dirname, '../node_modules/@zen/zen-node')
-}
-
-function isInstalledWithInstaller() {
-  // $FlowFixMe
-  return !process.resourcesPath.includes('node_modules/electron/dist')
-
-  // return __dirname.includes('app.asar') // tested on linux. below is an alternative
-  // TODO [AdGo] 15/05/2018 - delete these comments after confirming it works
-  // on os and windows
-}
 
 function getMainWindow(title) {
   const { width, height } = db.get('userPreferences').value()
@@ -153,38 +125,6 @@ function saveWindowDimensionsToDb() {
   if (!mainWindow) { return }
   const { width, height } = mainWindow.getBounds()
   db.get('userPreferences').assign({ width, height }).write()
-}
-
-function getZenNodeArgs() {
-  let args = []
-  if (process.env.WIPE || process.argv.indexOf('--wipe') > -1 || process.argv.indexOf('wipe') > -1) {
-    console.log('WIPING DB')
-    args.push('--wipe')
-  }
-  if (process.env.WIPEFULL || process.argv.indexOf('--wipe full') > -1 || process.argv.indexOf('wipefull') > -1) {
-    console.log('FULLY WIPING DB')
-    args = [...args, '--wipe', 'full']
-  }
-  if (process.env.MINER || process.argv.indexOf('--miner') > -1 || process.argv.indexOf('miner') > -1 || db.get('config.isMining').value()) {
-    console.log('RUNNING A MINER')
-    args.push('--miner')
-  }
-  if (process.env.ZEN_LOCAL) {
-    console.log('Running locally and mining')
-    args = [...args, '--chain', 'local', '--miner']
-  }
-  console.log('Zen node args', args)
-  return args
-}
-
-function setupLogFetching() {
-  ipcMain.on('init-fetch-logs', (event) => {
-    node.stdout.on('data', (chunk) => {
-      const log = chunk.toString('utf8')
-      console.log(`Received ${log} bytes of data.`)
-      event.sender.send('blockchainLogs', log)
-    })
-  })
 }
 
 function buildElectronMenu() {
